requirements.txt
colorama==0.4.4
requests==2.27.1
paramiko==2.9.2
pycryptodome==3.14.1